+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Add Croatia to LocalDetailsCleaner and IBANAssembler
 
 ## 0.6.5 - July 30, 2015
 - Add Romania to LocalDetailsCleaner and IBANAssembler